CSS changes
Mainly CSS chages to adapt the img button that will be used as a link as in the previous project. Failed attempts to change the backrgound color of the links led to an inconsistency in the code, but there are options that have not been explored yet, such as javascript usage. Link javascript and elaborate with  further css changes.