All pages finished and adjusted.
I kinda forgot to commit, so... here it goes. The last pages (fighting style and sabers have been finished, as well, as login, signup and recovery) for this project. None of them optimized yet and a few of them might break with slight changes,, since there is no media query to make it work yet. All pages linked to the respective paths and effects fixed (a few of them were not working well, due to the <a href> in a few others.

Please, don't mess with the code, if you don't know what you are doing.